requirements.txt
amqp==2.5.2
astroid==2.3.3
autopep8==1.4.4
backcall==0.1.0
billiard==3.6.1.0
celery==4.4.0
certifi==2019.11.28
chardet==3.0.4
decorator==4.4.1
Django==2.2.7
django-debug-toolbar==2.1
djangorestframework==3.11.0
idna==2.8
importlib-metadata==1.3.0
ipython==7.10.2
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.15.1
kombu==4.6.7
lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==8.0.2
parso==0.5.2
pexpect==4.7.0
pickleshare==0.7.5
pkg-resources==0.0.0
prompt-toolkit==3.0.2
ptyprocess==0.6.0
pycodestyle==2.5.0
Pygments==2.5.2
pylint==2.4.4
pytz==2019.3
redis==3.3.11
requests==2.22.0
sentry-sdk==0.13.5
six==1.13.0
sqlparse==0.3.0
traitlets==4.3.3
typed-ast==1.4.0
urllib3==1.25.7
vine==1.3.0
wcwidth==0.1.7
wrapt==1.11.2
zipp==0.6.0
